+++ Soggins/Cards/ExtraApologyCard.cs
@@ -49,16 +49,6 @@ public sealed class ExtraApologyCard : Card, IRegisterableCard
 				statusAmount = 1,
 				targetPlayer = true
 			},
-			ModEntry.Instance.KokoroApi.SpoofedActions.MakeAction(
-				new AAddCard
-				{
-					card = new RandomPlaceholderApologyCard(),
-					destination = CardDestination.Hand
-				},
-				new AAddApologyCard
-				{
-					Destination = CardDestination.Hand,
-				}
-			).AsCardAction,
+			new AAddApologyCard(),
 		];
 }