+++ Soggins/docs/release-notes.md
@@ -5,6 +5,7 @@
 ## Upcoming release
 
 * Fixed Smug-related dialogue not being properly serialized, which normally wouldn't really be noticeable, unless you're playing with the Combat QoL mod.
+* Apology-generating cards now use their own unique action.
 
 ## 1.8.0
 Released 22 June 2025.